fix
se corrige user controler para usar el dto
se modifica el delete del loan para que devuelva string
se corrige el getAllAccount para que los obtenga del repository

+++ src/main/java/com/codoacodo23650/tpgrupo14/services/LoanService.java
@@ -48,7 +48,14 @@ public class LoanService {
         return LoanMapper.loanToDto( repository.save(existingLoan));
     }
 
-    public void deleteLoan(Long loanId) {
-        repository.deleteById(loanId);
+    public String deleteLoan(Long loanId) {
+        if (repository.existsById(loanId)){
+            repository.deleteById(loanId);
+            return "El préstamo con id: " + loanId + " ha sido eliminada";
+        } else {
+            return "El préstamo con id: " + loanId + ", no ha sido eliminada";
+        }
+
+
     }
 }
